continuing work on the templating feature
